# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ePKL Backend API",
        "description": "Backend for ePKL mobile app",
        "contact": {
            "name": "Muhamamd Haikal",
            "email": "realityinaship@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/tugas": {
            "get": {
                "tags": [
                    "Tugas"
                ],
                "summary": "Get all Tugas",
                "description": "Retrieve a list of all Tugas.",
                "operationId": "getAllTugas",
                "responses": {
                    "200": {
                        "description": "Tugas retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tugas"
                ],
                "summary": "Create a Tugas",
                "description": "Create a new Tugas.",
                "operationId": "createTugas",
                "requestBody": {
                    "description": "Tugas details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "users_id": {
                                        "type": "integer"
                                    },
                                    "nama_tugas": {
                                        "type": "string"
                                    },
                                    "prioritas": {
                                        "type": "string"
                                    },
                                    "deadline": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tugas created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/tugas/{id}": {
            "get": {
                "tags": [
                    "Tugas"
                ],
                "summary": "Get a Tugas by ID",
                "description": "Retrieve a specific Tugas by its ID.",
                "operationId": "getTugas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Tugas",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tugas retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tugas"
                ],
                "summary": "Update a Tugas",
                "description": "Update an existing Tugas.",
                "operationId": "updateTugas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Tugas",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Tugas details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_tugas": {
                                        "type": "string"
                                    },
                                    "prioritas": {
                                        "type": "string"
                                    },
                                    "is_checked": {
                                        "type": "integer"
                                    },
                                    "deadline": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Tugas updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tugas"
                ],
                "summary": "Delete a Tugas",
                "description": "Delete a Tugas by its ID.",
                "operationId": "deleteTugas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Tugas",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tugas deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}